+++ src/entrypoints/api.ts
@@ -3,7 +3,7 @@ import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger'
 import { ValidationPipe, VersioningType } from '@nestjs/common'
 import { AppModule } from '../app.module.js'
 import { initSentry } from '../helpers/sentry.js'
-import { HttpExceptionFilter } from '../utils/exceptions/http-exception.filter.js'
+import { HttpExceptionFilter } from '../utils/exception/http-exception.filter.js'
 
 async function bootstrap (): Promise<void> {
   const app = await NestFactory.create(
