+++ test/setup/setup.ts
@@ -5,7 +5,7 @@ import { HttpAdapterHost } from '@nestjs/core'
 import { Test, type TestingModule } from '@nestjs/testing'
 import { expect } from 'expect'
 import { AppModule } from '../../src/app.module.js'
-import { HttpExceptionFilter } from '../../src/utils/exceptions/http-exception.filter.js'
+import { HttpExceptionFilter } from '../../src/utils/exception/http-exception.filter.js'
 import { uuid } from '../expect/expectUuid.js'
 import { toHaveErrorCode } from '../expect/expectErrorCode.js'
 import { toHaveStatus } from '../expect/expectStatus.js'
